EntityData.hp 타입 변경, Projectile 추상화 및 데이터 SO로 분리
EntityData.hp int -> float으로 변경

Projectile <- Straight 형태로 상속
현재는 단순히 이동 로직만 override하는 형태

Projectile의 데이터 damage, speed, range, duration을 SO로 구조화하여 SO폴더에서 따로 저장.

+++ Assets/Scripts/DamagePopup/DamagePopup.cs
@@ -16,7 +16,7 @@ public class DamagePopup : MonoBehaviour
     private float fadeDuration = 0.5f;
     private Vector3 offset;
 
-    public void Setup(Color color, int amount, Vector3 point, float duration, DamagePopupManager damagePopupManager)
+    public void Setup(Color color, float amount, Vector3 point, float duration, DamagePopupManager damagePopupManager)
     {
         damageText.color = color;
         damageText.text = amount.ToString();
