+++ Assets/Scripts/IDamageable.cs
@@ -3,7 +3,7 @@ using UnityEngine.Events;
 
 public interface IDamageable
 {
-    public void TakeDamage(int amount, Entity enemy, string weaponName);
+    public void TakeDamage(float amount, Entity enemy, string weaponName);
 
     public bool IsDead { get;}
 }
